+++ src/models/object.ts
@@ -1,0 +1,28 @@
+export interface IObject {
+    type: number;
+    id: string;
+    enemy: boolean;
+    item: boolean;
+    god: boolean;
+    pet: boolean;
+    slotType: number;
+    bagType: number;
+    class: string;
+    maxHitPoints: number;
+    defense: number;
+    xpMultiplier: number;
+    projectile: {
+        id: number;
+        objectId: string;
+        damage: number;
+        minDamage: number;
+        maxDamage: number;
+        speed: number;
+        lifetimeMS: number;
+    };
+    rateOfFire: number;
+    fameBonus: number;
+    feedPower: number;
+    occupySquare: boolean;
+    fullOccupy: boolean;
+}